fix: resolve POST redirect issue and improve monster creation
- Fix Nginx configuration to handle POST requests without trailing slash
- Change Danger field from int to double for fractional values (0.25, 0.5, etc.)
- Add unique constraint for monster names per user
- Implement direct DB access for usernames via GatewayDbContext
- Add CreatedByUsername field to Monster model
- All CRUD operations now working correctly"

+++ SheetService/DTO/MonsterDTO.cs
@@ -12,7 +12,7 @@ namespace SheetService.DTOs
         public int Int { get; set; }
         public int Wis { get; set; }
         public int Cha { get; set; }
-        public int Danger { get; set; }
+        public double Danger { get; set; }
         public string Description { get; set; } = string.Empty;
         public string CreatedBy { get; set; } = string.Empty;
         public string CreatedByUsername { get; set; } = string.Empty;
@@ -32,7 +32,7 @@ namespace SheetService.DTOs
         public int Int { get; set; }
         public int Wis { get; set; }
         public int Cha { get; set; }
-        public int Danger { get; set; }
+        public double Danger { get; set; }
         public string Description { get; set; } = string.Empty;
         public string Status { get; set; } = "private";
     }
